Cleaning up the app, moving built-out of the way so we can build more
