+++ README.md
@@ -35,8 +35,4 @@ npm install @tailwindcss/typography
 
 ng update @angular/core@16 @angular/cli@16
 
-TODO: JAY
 
-And RP fix 
-
-Hello Michael
