+++ src/app/app-routing.module.ts
@@ -1,19 +1,19 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes,  TitleStrategy, RouterStateSnapshot } from '@angular/router';
 // https://www.telerik.com/blogs/angular-basics-router-link-overview
-import { GuestProfileComponent } from './guest-profile/guest-profile.component';
+import { GuestProfileComponent } from './built/guest-profile/guest-profile.component';
 import { SitemapComponent } from './sitemap/sitemap.component';
 import { HeaderComponent } from './header/header.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 import { A01c2GuestAboutComponent } from './a01c2-guest-about/a01c2-guest-about.component'
 import { A01c4GuestWebLinksComponent } from './a01c4-guest-web-links/a01c4-guest-web-links.component'
-import { A02aUpdateYourPasswordComponent } from './a02a-update-your-password/a02a-update-your-password.component'
+// import { A02aUpdateYourPasswordComponent } from './built/a02a-update-your-password/a02a-update-your-password.component'
 
 import { A01aGuestSignInComponent } from './a01a-guest-sign-in/a01a-guest-sign-in.component';
 import { A01aGuestSignUpComponent } from './a01a-guest-sign-up/a01a-guest-sign-up.component';
-import { A04aHamburgerNavComponent } from './a04a-hamburger-nav/a04a-hamburger-nav.component';
+// import { A04aHamburgerNavComponent } from './built/a04a-hamburger-nav/a04a-hamburger-nav.component';
 import { A05bGuestPublicLandingComponent } from './a05b-guest-public-landing/a05b-guest-public-landing.component';
-import { B04aVendorAboutLandingDisplayComponent } from './b04a-vendor-about-landing-display/b04a-vendor-about-landing-display.component';
+// import { B04aVendorAboutLandingDisplayComponent } from './built/b04a-vendor-about-landing-display/b04a-vendor-about-landing-display.component';
 import { FooterComponent } from './footer/footer.component';
 
 
@@ -25,9 +25,9 @@ const routes: Routes = [
   { path: 'guest-sign-in', component: A01aGuestSignInComponent},
   { path: 'guest-sign-up', component: A01aGuestSignUpComponent},
   
-  { path: 'update-password', component: A02aUpdateYourPasswordComponent},
-  { path: 'vendor-landing', component: B04aVendorAboutLandingDisplayComponent},
-  { path: 'hamburger-nav', component: A04aHamburgerNavComponent},
+  // { path: 'update-password', component: A02aUpdateYourPasswordComponent},
+  // { path: 'vendor-landing', component: B04aVendorAboutLandingDisplayComponent},
+  // { path: 'hamburger-nav', component: A04aHamburgerNavComponent},
   { path: 'sitemap', component: SitemapComponent },
   { path: 'header', component: HeaderComponent },
   { path: 'footer', component: FooterComponent },
