+++ src/app/built/a04a-hamburger-nav/a04a-hamburger-nav.component.html
@@ -1,7 +1,5 @@
 <div class="px-6 pt-10 flex justify-between">
-
     <div>
-
         <div class="flex">
             <div class="rounded-full bg-gray-400">
                 <img class=" p-6"
@@ -20,7 +18,6 @@
                 </div>
             </div>
         </div>
-
     </div>
 
     <div class="flex h-fit w-40 justify-between mr-2">
@@ -64,7 +61,6 @@
         <span class="text-2xl ml-4">></span>
     </p>
 </div>
-
 
 <div class="flex w-screen border-b-2 pb-6 mt-4 border-gray-400">
     <p class="text-md ml-8 font-light">
