+++ src/app/built/b04a-vendor-about-landing-display/b04a-vendor-about-landing-display.component.ts
