+++ src/app/header/header.component.html
@@ -1,22 +1,14 @@
 <header class="bg-slate-200 border-stone-700 border-2">
 
- 
   <mat-accordion>
     <mat-expansion-panel hideToggle></mat-expansion-panel>
-  
-    <mat-expansion-panel (opened)="panelOpenState = true"
-                         (closed)="panelOpenState = false">
+    <mat-expansion-panel (opened)="panelOpenState = true" (closed)="panelOpenState = false">
       <mat-expansion-panel-header>
-        <mat-panel-title>
-          Menu
-        </mat-panel-title>
-        <mat-panel-description>
-          Currently I am {{panelOpenState ? 'open' : 'closed'}}
-        </mat-panel-description>
+        <mat-panel-title>Menu</mat-panel-title>
+        <mat-panel-description>Currently I am {{panelOpenState ? 'open' : 'closed'}}</mat-panel-description>
       </mat-expansion-panel-header>
     
-      <nav>
-        <ul>
+      <nav><ul>
           <li><a routerLink="/guest" routerLinkActive="active" ariaCurrentWhenActive="page" >guest</a></li>
           <li><a routerLink="/guest-about" routerLinkActive="active" ariaCurrentWhenActive="page" >guest-about</a></li>
           <li><a routerLink="/guest-landing" routerLinkActive="active" ariaCurrentWhenActive="page" >guest-landing</a></li>
@@ -24,18 +16,13 @@
           <li><a routerLink="/guest-sign-in" routerLinkActive="active" ariaCurrentWhenActive="page" >guest-sign-in</a></li>
           <li><a routerLink="/guest-sign-up" routerLinkActive="active" ariaCurrentWhenActive="page" >guest-sign-up</a></li>
     
-          <li><a routerLink="/update-password" routerLinkActive="active" ariaCurrentWhenActive="page" >update-password</a></li>
-          <li><a routerLink="/vendor-landing" routerLinkActive="active" ariaCurrentWhenActive="page" >vendor-landing</a></li>
-          <li><a routerLink="/hamburger-nav" routerLinkActive="active" ariaCurrentWhenActive="page" >hamburger-nav</a></li>
+          <!-- <li><a routerLink="/update-password" routerLinkActive="active" ariaCurrentWhenActive="page" >update-password</a></li> -->
+          <!-- <li><a routerLink="/vendor-landing" routerLinkActive="active" ariaCurrentWhenActive="page" >vendor-landing</a></li> -->
+          <!-- <li><a routerLink="/hamburger-nav" routerLinkActive="active" ariaCurrentWhenActive="page" >hamburger-nav</a></li> -->
           <li><a routerLink="/sitemap" routerLinkActive="active" ariaCurrentWhenActive="page" >sitemap</a></li>
           <li><a routerLink="/page-not-found" routerLinkActive="active" ariaCurrentWhenActive="page" >page-not-found</a></li>   
-    
-          <!-- <li><a routerLink="/foo" routerLinkActive="active" ariaCurrentWhenActive="page" >foo</a></li>
-          <li><a routerLink="/boo" routerLinkActive="active" ariaCurrentWhenActive="page" >boo</a></li> -->
-          
         </ul>
       </nav>
     </mat-expansion-panel>
   </mat-accordion>
-
 </header>
